+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # CHANGELOG
 
+## 1.14.0
+### 05-Feb-2024
+- Updated Vimeo ID
+- Corrected omission of [SYM-Helper] for `moveableInProduction`
+- Updated "Microsoft Office 365" to "Microsoft 365"
+- Added OS Build number to webhook output [Pull Request No. 124](https://github.com/dan-snelson/Setup-Your-Mac/pull/124); thanks, @drtaru!
+- Changed filepath validation test from `-f` (i.e., "True if file exists and is a regular file") to `-e` (i.e., "True if file exists (regardless of type)."); thanks for the inspiration, @mrmte! [Issue 19](https://github.com/BIG-RAT/SYM-Helper/issues/19); thanks for the code suggestion, @bartreardon!
+- Updates to `README.md`, `CONTRIBUTORS.md` and `CONTRIBUTING.md` [Pull Request No. 128](https://github.com/setup-your-mac/Setup-Your-Mac/pull/128); thanks, @robjschroeder!
+- Refactored the way `brandingBanner` variable is checked [Pull Request No. 131](https://github.com/setup-your-mac/Setup-Your-Mac/pull/131); thanks, @drtaru!
+- Increased minimum required version of swiftDialog to `2.4.0.4750`
+- Leveraged the new `listitem: subtitle` option with a dedicated `subtitle` field in `policyJSON`
+- Corrected misspelling of "policies" in log entries [Issue No. 134](https://github.com/setup-your-mac/Setup-Your-Mac/issues/134); thanks, @Honestpuck!
+- Updated `brandingBanner` to [image by benzoix on Freepik](https://www.freepik.com/author/benzoix)
+
 ## 1.13.0
 ### 24-Oct-2023
 - 🔥 **Breaking Change** for users of Setup Your Mac prior to `1.13.0` 🔥